Sample ancestral population with ms 2 island model. Then admixe ancestral population

Sample ancestral population with ms 2 island model. Then admixe ancestral population along a longitudinal gradient.

SampleGenoOFWithMs(n, nsites.neutral, nsites.selected, crossover.proba,
  m.neutral, m.selected, mutation.rate.per.site, N0 = 10^6, k = 0.5,
  min.maf = 0.05, plot.debug = FALSE, tess3.ms = getOption("tess3.ms"))

Arguments

n

number of indivudual to sample

nsites.neutral

number of site between which recombination occur for neutral loci

nsites.selected

number of site between which recombination occur for selected loci

crossover.proba

corss-over probability between adjacent site per generation

m.neutral

migration rate for neutral loci

m.selected

migration rate for selected loci

mutation.rate.per.site

mutation rate per site

N0

population size

k

TODOC

min.maf

the locus with a maf less than this parameter are removed

plot.debug

if TRUE plot at different stage of the simulation

tess3.ms

ms binary path.

Value

TODOC

Examples

# tess3.ms <- "~/BiocompSoftware/msdir/ms"
# n <- 200
# K <- 2
# ploidy <- 1
# data.list <- SampleGenoOFWithMs(n = n,
#                              nsites.neutral = 100000,
#                              nsites.selected = 1000,
#                              crossover.proba = 0.25 * 10 ^ -8,
#                              m.neutral = 0.25 * 10 ^ -6,
#                              m.selected = 0.25 * 10 ^ -7,
#                              mutation.rate.per.site = 0.25 * 10 ^ -8,
#                              N0 = 10 ^ 6,
#                              k = 0.5,
#                              min.maf = 0.05,
#                              plot.debug = TRUE)