Estimates spatial population structure

tess3Main estimates spatial population structure using a graph based non negative matrix factorization. After estimating the population structure is used to compute a Fst statistic for each locus. See references for more details.

tess3Main(X, XProba = NULL, coord, K, ploidy, lambda = 1, W = NULL,
  method = "projected.ls", max.iteration = 200, tolerance = 1e-05,
  openMP.core.num = 1, Q.init = NULL, mask = 0, copy = TRUE,
  algo.copy = TRUE, verbose = FALSE)

Arguments

X

A numeric matrix which corresponds to the genotype matrix. This matrix must be of size \(n \times L\) where \(n\) is the number of individual and \(L\) is the number of loci. Values of this matrix are integer corresponding to the number of variant alleles observed at a locus. If NULL, XProba is used.

XProba

A numeric matrix which correspond to the probability for the genotype. This matrix must be of size \(n \times (ploidy + 1)L\) where \(n\) is the number of individual, \(L\) is the number of loci. Values of this matrix are numeric between 0 and 1 corresponding to the genome probability. It is the matrix used in graph based non negative factorization matrix. If NULL, it is computed from the genotype matrix X. See reference for more details.

coord

The numeric matrix of size \(n \times 2\) where \(n\) is the number of individuals.

K

An integer which corresponds to the number of ancestral populations.

ploidy

An integer which corresponds to the number of set of chromosomes.

lambda

A numeric which corresponds to the spatial regularization parameter.

W

A numeric matrix which corresponds to the graph weiht matrix. If NULL, W it is computed as W[i,j] = exp( - (coord[i] - coord[j])^2 / sigma^2). Where coord[i] is the geographic coordinate for the individual i and sigma equals 5 percent of the average geographic distance between individuals.

method

"projected.ls" or "qp". If "projected.ls", an aleternated projected least squares algorithm is used. If "qp", an alternated quadratic programing algorithm is used. See references for more details

max.iteration

The max number of iteration of the optimization algorithm.

tolerance

A numeric which corresponds to the tolerance of the stopping criteria of the optimization algorithm.

openMP.core.num

If openMP is available on your computer, it is the number of core used by the algorithm.

Q.init

A numeric matrix which corresponds to the initial value of Q for the algorithm.

mask

If not NULL this the proportion of the data matrix which is masked to compute the cross validation criteria.

copy

if TRUE data will be copy 1 time.

algo.copy

if TRUE data will be copy 1 time to speed the algorithm.

verbose

If TRUE more information are printed.

Value

An object of class tess3Main which is a list with components:

L

The number of loci.

n

The number of individuals.

ploidy

The number of set of chromosomes.

K

The number of ancestral population.

G

The ancestral genotype frequency matrix.

Q

The ancestry coefficient matrix.

Fst

The Fst statistic computed for each locus.

Fscore

The Fscore computed from Fst.

pvalue

The pvalues computed from Fscore.

log.pvalue

The \(log(pvalue)\).

rmse

The root square mean error between XProba and tcrossprod(Q, G).

crossentropy

The cross entropy error between XProba and tcrossprod(Q, G).

crossvalid.rmse

If masked not NULL. The root square mean error between XProba[masked] and tcrossprod(Q, G)[masked].

crossvalid.crossentropy

If masked not NULL. The cross entropy error between XProba[masked] and tcrossprod(Q, G)[masked].

References

https://hal.archives-ouvertes.fr/hal-01222555/ http://biorxiv.org/content/early/2016/10/12/080291

See also

tess3

Examples

library(tess3r)

# Arabidopsis thaliana data set
data(data.at)
genotype <- data.at$X
coordinates <- data.at$coord

# Run of tess3 main algorithm
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1)

#> Computing spectral decomposition of graph laplacian matrix: done
#> Main loop with 1 threads: 
#> : done
#> 

# Run of tess3 main algorithm with cross validation errors computation.
tess3.obj <- tess3Main(X = genotype,
                      coord = coordinates,
                      K = 3,
                      method = "projected.ls",
                      ploidy = 1,
                      mask = 0.05)

Mask 0.05% of genotypes for cross validation

Missing value detected in genotype

#> Computing spectral decomposition of graph laplacian matrix: done
#> Main loop with 1 threads: 
#> : done
#>